backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
transformers==4.41.2
torch==2.3.1
torchvision==0.18.1
requests==2.32.3
pandas==2.2.2
scikit-learn==1.5.1
numpy==1.26.4
Pillow==10.3.0
firebase-admin==6.5.0
kaggle==1.6.12
sentencepiece==0.2.0
accelerate==0.30.1
setuptools>=65.0.0
wheel>=0.41.0